% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs_metadata.R
\name{get_pkgs_metadata}
\alias{get_pkgs_metadata}
\title{List of metadata for all packages from a given R-universe}
\usage{
get_pkgs_metadata(universe = "epiverse-connect")
}
\arguments{
\item{universe}{Name of the r-universe to explore}
}
\value{
A list of list where each element refers to a specific package and
contains the following elements:
\itemize{
\item \code{Package} (character): package name
\item \code{logo} (character): Link to the package logo
\item \code{website} (character): Link to the documentation website
\item \code{source} (character): Link to the package source
\item \code{articles} (character vector): Links to the package vignettes
}
}
\description{
List of metadata for all packages from a given R-universe
}
