% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_universe_docs.R
\name{get_universe_docs}
\alias{get_universe_docs}
\title{Fetch documentation files for all packages from a given R-universe}
\usage{
get_universe_docs(universe = "epiverse-connect", destdir = "docs")
}
\arguments{
\item{universe}{Name of the r-universe to explore}

\item{destdir}{Path where the documentation files should be written (defaults
to \code{"docs"}).}
}
\value{
This function is called for its side effect. The documentation files
(individual function documentation under \verb{man/} and vignettes) are stored
on the file system, in the folder specified in \code{destdir}, with the specific
folder structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{docs
├── pkg1
│   └── man
│       ├── fun1.md
│       ├── fun2.md
│       └── fun3.md
└── pkg2
    ├── man
    │   ├── fct1.md
    │   ├── fct2.md
    └── vignettes
        ├── vig_a.Rmd
        ├── vig_b.Rmd
}\if{html}{\out{</div>}}
}
\description{
Fetch documentation files for all packages from a given R-universe
}
